+++ docs/conf.py
@@ -31,7 +31,7 @@ project = 'PySight'
 year = '2017'
 author = 'Hagai Hargil'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '0.5.19'
+version = release = '0.5.20'
 autodoc = True
 
 pygments_style = 'trac'